Bug 1731631 - use open bug for classification if duplicate one matches failure line
Bug A has been set as duplicate of bug B. The summary of bug A matches the
failure line. To aggregate all classifications against the common root cause,
classifications will automatically be store against the open bug B. If a bug got
set as duplicate, existing classifications for the duplicate will be updated to
use the open bug against which the duplicate has been set.